+++ gg/include/debug/DebugUtils.h
@@ -11,20 +11,20 @@
 
 namespace gg { namespace debug {
 	void drawLine(const math::Vec3f& from, const math::Vec3f& to, const math::Color& color = math::Color::white);
-	void drawLine(const math::Vec3f& from, const math::Vec3f& to, float thickness, const math::Color& color = math::Color::white);
-	void drawLine(const math::Vec3f& from, const math::Vec3f& to, float thickness, float duration, const math::Color& color = math::Color::white);
+	void drawLine(const math::Vec3f& from, const math::Vec3f& to, uint thickness, const math::Color& color = math::Color::white);
+	void drawLine(const math::Vec3f& from, const math::Vec3f& to, uint thickness, float duration, const math::Color& color = math::Color::white);
 
 	void drawWireCube(const math::Vec3f& origin, float size, const math::Color& color = math::Color::white);
-	void drawWireCube(const math::Vec3f& origin, float size, float thickness, const math::Color& color = math::Color::white);
-	void drawWireCube(const math::Vec3f& origin, float size, float thickness, float duration, const math::Color& color = math::Color::white);
+	void drawWireCube(const math::Vec3f& origin, float size, uint thickness, const math::Color& color = math::Color::white);
+	void drawWireCube(const math::Vec3f& origin, float size, uint thickness, float duration, const math::Color& color = math::Color::white);
 
 	void drawWireCube(const math::Vec3f& origin, float size, const math::Vec3f& euler, const math::Color& color = math::Color::white);
-	void drawWireCube(const math::Vec3f& origin, float size, const math::Vec3f& euler, float thickness, const math::Color& color = math::Color::white);
-	void drawWireCube(const math::Vec3f& origin, float size, const math::Vec3f& euler, float thickness, float duration, const math::Color& color = math::Color::white);
+	void drawWireCube(const math::Vec3f& origin, float size, const math::Vec3f& euler, uint thickness, const math::Color& color = math::Color::white);
+	void drawWireCube(const math::Vec3f& origin, float size, const math::Vec3f& euler, uint thickness, float duration, const math::Color& color = math::Color::white);
 
 	void drawWireCube(const math::Vec3f& origin, float size, const math::Quaternion& rot, const math::Color& color = math::Color::white);
-	void drawWireCube(const math::Vec3f& origin, float size, const math::Quaternion& rot, float thickness, const math::Color& color = math::Color::white);
-	void drawWireCube(const math::Vec3f& origin, float size, const math::Quaternion& rot, float thickness, float duration, const math::Color& color = math::Color::white);
+	void drawWireCube(const math::Vec3f& origin, float size, const math::Quaternion& rot, uint thickness, const math::Color& color = math::Color::white);
+	void drawWireCube(const math::Vec3f& origin, float size, const math::Quaternion& rot, uint thickness, float duration, const math::Color& color = math::Color::white);
 }/*namespace debug*/ } // namespace gg
 
 #endif
