+++ gg/src/graphics/Texture.cpp
@@ -83,7 +83,7 @@ namespace gg { namespace graphics {
 
 	math::Vec2f Texture::getSize(void) const
 	{
-		return math::Vec2f(m_Width, m_Height);
+		return math::Vec2f((float)m_Width, (float)m_Height);
 	}
 	int Texture::getWidth(void) const
 	{
